refactor(sig-events): revert sig events tab changes
