+++ x-pack/platform/plugins/shared/streams/server/routes/internal/sig_events/verdicts/route.ts
@@ -69,7 +69,7 @@ const verdictsHistoryRoute = createServerRoute({
 
     await assertSignificantEventsAccess({ server, licensing, uiSettingsClient });
 
-    return getVerdictClient().findByDiscoverySlug(params.path.discoveryId);
+    return getVerdictClient().findByDiscoveryId(params.path.discoveryId);
   },
 });
