+++ x-pack/platform/plugins/shared/streams_app/public/components/sig_events/significant_events_discovery/page.tsx
@@ -30,6 +30,7 @@ import { SettingsTab } from './components/settings/tab';
 import { MemoryTab } from './components/memory/tab';
 import { DetectionsTab } from './components/detections_tab';
 import { DiscoveriesTab } from './components/discoveries_tab';
+import { VerdictsTab } from './components/verdicts_tab';
 import { SigEventsTab } from './components/sig_events_tab';
 
 const discoveryTabs = [
@@ -40,6 +41,7 @@ const discoveryTabs = [
   'detections',
   'discoveries',
   'verdicts',
+  'significant_events',
   'memory',
   'settings',
 ] as const;
@@ -135,10 +137,18 @@ export function SignificantEventsDiscoveryPage() {
       {
         id: 'verdicts',
         label: i18n.translate('xpack.streams.significantEventsDiscovery.verdictsTab', {
-          defaultMessage: 'Significant Events',
+          defaultMessage: 'Verdicts',
         }),
         href: router.link('/_discovery/{tab}', { path: { tab: 'verdicts' } }),
         isSelected: tab === 'verdicts',
+      },
+      {
+        id: 'significant_events',
+        label: i18n.translate('xpack.streams.significantEventsDiscovery.sigEventsTab', {
+          defaultMessage: 'Significant Events',
+        }),
+        href: router.link('/_discovery/{tab}', { path: { tab: 'significant_events' } }),
+        isSelected: tab === 'significant_events',
       },
       {
         id: 'insights',
@@ -234,7 +244,8 @@ export function SignificantEventsDiscoveryPage() {
           {tab === 'insights' && <InsightsTab />}
           {tab === 'detections' && <DetectionsTab />}
           {tab === 'discoveries' && <DiscoveriesTab />}
-          {tab === 'verdicts' && <SigEventsTab />}
+          {tab === 'verdicts' && <VerdictsTab />}
+          {tab === 'significant_events' && <SigEventsTab />}
           {tab === 'memory' && isMemoryEnabled && <MemoryTab />}
           {tab === 'settings' && <SettingsTab />}
         </StreamsAppPageTemplate.Body>
